+++ test/data.cmake
@@ -7,19 +7,19 @@
 odr_test_data(
         PATH "input/odr-public"
         URL "https://github.com/opendocument-app/OpenDocument.test.git"
-        REVISION "d7c093dffc987e8e5cf093ea7412371ce9d7363d")
+        REVISION "77ed98d72075bbd375be328b7232f40568a32563")
 
 odr_test_data(
         PATH "input/odr-private"
         URL "https://github.com/opendocument-app/OpenDocument.test-private.git"
-        REVISION "753b7e5c313c23c0524cb9ed678ac294508f3e21")
+        REVISION "d92bbbc453e6dbc0187ec2bdf0560a6e48d5643b")
 
 odr_test_data(
         PATH "reference-output/odr-public"
         URL "https://github.com/opendocument-app/OpenDocument.test.output.git"
-        REVISION "1964c34c2a6bd84d8178d9586ba59a04f172b7f2")
+        REVISION "6d1ed797b3833b99b09ce359f2689c5a3af105c1")
 
 odr_test_data(
         PATH "reference-output/odr-private"
         URL "https://github.com/opendocument-app/OpenDocument.test-private.output.git"
-        REVISION "5e5f035f8a6e6c86ad5d978887607e8b9167c019")
+        REVISION "90841441c293ec665632b570bdfe63842926859c")